refactor: refactoring all components created in `header` component"
Separate each component part in each of those component files
